+++ src/ResourceManager/AnalysisServices/ChangeLog.md
@@ -18,9 +18,9 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Added two new dataplane APIs in a separate module AzureRM.AnalysisServices.Dataplane.psd1
+    - This introduces two new APIs that enable customers to login to Azure Analysis Services servers and issue a restart command.
 
-## Version 3.3.0
+## Version 3.2.0
 * Fixed bug in Get-AzureRMAnalysisServicesServer
     - When this command was run against some resources, it would fail with a null reference exception.
-* Added two new dataplane APIs in a separate module AzureRM.AnalysisServices.ServiceManagement.psd1
-    - This introduces two new APIs that enable customers to login to Azure Analysis Services servers and issue a restart command.